vastai-deployment/requirements.txt
vastai>=0.2.8
python-dotenv>=1.0.0
colorama>=0.4.6 